third_party/WebKit/LayoutTests/external/wpt/WebIDL/ecmascript-binding/interface-prototype-object-expected.txt

This is a testharness.js-based test.
FAIL The class string of an interface prototype object is the concatenation of the interface's identifier and the string 'Prototype'. assert_equals: expected "[object DocumentPrototype]" but got "[object Document]"
Harness: the test ran to completion.
